Increment System.Threading.Channels to latest version number
This causes conflicts in programs if you do not specify System.Threading.Channels as a dependency, and you compile with the runtime flag to produce a self-contained executable. If you manually include the dependency in the main project at the lower version everything is fine, this just makes it easier for everyone as this new version does not appear to break anything.