CI: try to deal with testing framework
I didn't even notice that there _was_ a testing framework, which uses a 'special' SL account to log in, etc. Sadly, the way the login information is passed to GitHub Actions is anything but obvious; as such, this will be my last experiment — I'll skip the testing part if this doesn't work. After all, AppVeyor will run all the tests properly...